moved sending email to notification class
