+++ zebrand-backend/app/product/use_cases/product_use_case.py
@@ -1,14 +1,7 @@
 import math
 from datetime import datetime
-from pathlib import Path
 
-from fastapi import Depends, status, Request
-from fastapi.security import HTTPBearer
-
-from jose import jwt
-
-from jinja2 import Template
-import boto3
+from fastapi import Depends, status
 
 from app.product.services.product_service import ProductService
 from app.product.domain.responses.product_response import ProductResponse, GetProductResponse, PaginationMetadata, ProductChanges
@@ -16,8 +9,6 @@ from app.product.domain.requests.product_request import ProductRequest, UpdatePr
 from app.product.domain.models.product_model import ProductModel
 from app.product.domain.models.product_track_view_model import ProductTrackViewModel
 from app.product.domain.exceptions.product_exception import ProductException
-from app.domain.settings.security_settings import get_security_setting
-from app.domain.settings.aws_settings import get_ses_client
 
 
 class ProductUseCase:
@@ -51,8 +42,6 @@ class ProductUseCase:
             )
             for product in products
         ]
-
-        print("PRODUCTS", product_response)
 
         response = GetProductResponse(
             metadata=pagination,
@@ -132,61 +121,3 @@ class ProductUseCase:
     def track_one_product_visit(self, product: ProductResponse, keyword: str) -> None:
         tracked_product = ProductTrackViewModel(sku=product.sku, keyword=keyword)
         self._service.save_product_visited(tracked_product)
-
-    async def get_author_from_request(self, request: Request):
-        security = HTTPBearer(auto_error=False)
-        credentials = await security(request)
-            
-        security_settings = get_security_setting()
-
-        payload = jwt.decode(
-            credentials.credentials,
-            security_settings.secret_key,
-            algorithms=[security_settings.algorithm]
-        )
-
-        author = payload["sub"]
-        return author
-    
-    def get_admin_email_list(self) -> list[str]:
-        admins = self._service.get_all_admin_user()
-        emails = [admin.email for admin in admins]
-        return emails
-
-    async def send_product_change_email(self, request: Request, product_changes: list[ProductChanges]):
-        author = await self.get_author_from_request(request)
-        emails = self.get_admin_email_list()
-        
-        changes = [change.model_dump() for change in product_changes]
-        datenow = datetime.now()
-
-        path = Path(__file__).parent.parent
-        template_path = str(path / "domain" / "templates" / "product_change_template.html")
-        template = Template(template_path)
-
-        client = get_ses_client()
-
-        html_content = template.render(
-            author=author,
-            changes=changes,
-            current_date=datenow
-        )
-        response = client.send_email(
-            Destination={
-                'ToAddresses': emails,
-            },
-            Message={
-                'Body': {
-                    'Html': {
-                        'Charset': "UTF-8",
-                        'Data': html_content,
-                    },
-                },
-                'Subject': {
-                    'Charset': "UTF-8",
-                    'Data': "Product Update Notification",
-                },
-            },
-            Source=author,
-        )
-        print(response)
